IT->BLA is now modified to take the US amplitude into account.
The learing rule is a bit adhoc, as it aims at ensuring the sum(mod) is equal to sum(exc).

BLA now does not respond to the US, perhaps better?

Response to the US is now too big before learning, as CE and LH both excite PPTN. Make PPTN activation dependent on the combination of both info? eg CE->PPTN is modulatory (as IT->BLA).

The CS increases still too fast, the the US is not totally cancelled.